requirements.txt
streamlit
chromadb
sentence-transformers
openai
python-docx